+++ Localize/lang/strings.json
@@ -265,6 +265,7 @@
   "63fQWE": "Show all advanced parameters",
   "6DZp5H": "Search",
   "6ELsbA": "Profile",
+  "6F4Knf": "Guideline for agent advanced parameters",
   "6LJZ7n": "Retry policy",
   "6OSgRP": "Test map",
   "6PdOcy": "Cancel",
@@ -1465,6 +1466,7 @@
   "_63fQWE.comment": "Button tooltip to add all advanced parameters",
   "_6DZp5H.comment": "Placeholder text for search connectors",
   "_6ELsbA.comment": "The tab label for the monitoring profile tab on the configure template wizard",
+  "_6F4Knf.comment": "Agent advance parameters description link",
   "_6LJZ7n.comment": "title for retry policy setting",
   "_6OSgRP.comment": "Test map panel header",
   "_6PdOcy.comment": "Cancel",
@@ -2464,7 +2466,6 @@
   "_c8UPLp.comment": "Section 2 of text for including dynamic content section",
   "_cBw7SC.comment": "Label for connection creation date",
   "_cC7hlj.comment": "Agent Instruction Editor description link",
-  "_cC7prj.comment": "Agent advance parameters description link",
   "_cHiBAn.comment": "Time zone value ",
   "_cJkSrD.comment": "tooltip text of pagination setting",
   "_cKNvk6.comment": "Label for Value",
@@ -2619,7 +2620,6 @@
   "_gusZw5.comment": "Duplicate output names",
   "_gvDMuq.comment": "Select a Batch Workflow resource",
   "_gwEKLM.comment": "This is a message shown while loading. This announced text is read aloud with screen readers. Not shown in text.",
-  "_gzw2sh.comment": "Agent advance parameters description",
   "_gzwIqh.comment": "Agent Instruction Editor description",
   "_h+W3VW.comment": "Label for Number type dynamically added parameter",
   "_h+ZYip.comment": "Option to install a new gateway, links to new page",
@@ -3206,6 +3206,7 @@
   "_yBrsFJ.comment": "Required string parameter to be converted using array function",
   "_yENrOg.comment": "This is the link text in the message bar",
   "_yF2R//.comment": "Label for description of custom substring Function",
+  "_yJfhNL.comment": "Agent advance parameters description",
   "_yKNKV/.comment": "Accessbility text to indicate to try different search term or remove filters",
   "_yKOsmK.comment": "Label for description of custom sort Function",
   "_yNtBUV.comment": "Warning message for when input node type does not match one of the function node input's allowed types",
@@ -3326,7 +3327,6 @@
   "c8UPLp": "Dynamic content may also be added from other sources.",
   "cBw7SC": "Created",
   "cC7hlj": "Tips for writing agent instructions",
-  "cC7prj": "Guideline for agent advanced parameters",
   "cHiBAn": "(UTC+09:00) Seoul",
   "cJkSrD": "Retrieve more results up to the pagination limit",
   "cKNvk6": "{label} value item",
@@ -3481,7 +3481,6 @@
   "gusZw5": "Output names should be unique.",
   "gvDMuq": "Select a Batch Workflow resource",
   "gwEKLM": "Loading...",
-  "gzw2sh": "Add advanced parameters to further customize how the language model responds. These advanced options are optional and can help you fine-tune the experience.",
   "gzwIqh": "Add instructions so the agent understands its role and tasks. It can be helpful to include information about workflow structure, restrictions, tools, and interaction in certain scenarios.",
   "h+W3VW": "Number",
   "h+ZYip": "{addIcon} Install gateway",
@@ -4068,6 +4067,7 @@
   "yBrsFJ": "Required. The value that is converted to an array.",
   "yENrOg": "here",
   "yF2R//": "Returns a subset of characters from a string.",
+  "yJfhNL": "Add advanced parameters to further customize how the language model responds. These advanced options are optional and can help you fine-tune the experience.",
   "yKNKV/": "Try a different search term or remove filters",
   "yKOsmK": "Returns an array sorted in ascending order",
   "yNtBUV": "Warning: input node type does not match one of the allowed types for this input.",
