+++ libs/designer/src/lib/ui/settings/settingsection.tsx
@@ -8,7 +8,7 @@ import { CustomizableMessageBar } from './validation/errorbar';
 import type { ValidationError } from './validation/validation';
 import { ValidationErrorType } from './validation/validation';
 import type { IDropdownOption } from '@fluentui/react';
-import { Button, Divider, type MessageBarIntent, Tooltip, Badge, Link, Text } from '@fluentui/react-components';
+import { Button, Divider, type MessageBarIntent, Tooltip, Badge } from '@fluentui/react-components';
 import {
   bundleIcon,
   ChevronDown24Filled,
@@ -17,8 +17,6 @@ import {
   ChevronRight24Regular,
   Dismiss24Filled,
   Dismiss24Regular,
-  LinkSquare12Regular,
-  LinkSquare12Filled,
 } from '@fluentui/react-icons';
 import {
   MultiSelectSetting,
@@ -58,8 +56,8 @@ import { useIntl } from 'react-intl';
 import { useDispatch } from 'react-redux';
 import { useActionMetadata } from '../../core/state/workflow/workflowSelectors';
 import constants from '../../common/constants';
+import { AdvancedSettingsMessage } from './advancedSettingsMessage';
 
-const NavigateIcon = bundleIcon(LinkSquare12Regular, LinkSquare12Filled);
 const ClearIcon = bundleIcon(Dismiss24Filled, Dismiss24Regular);
 const ChevronDownIcon = bundleIcon(ChevronDown24Filled, ChevronDown24Regular);
 const ChevronRightIcon = bundleIcon(ChevronRight24Filled, ChevronRight24Regular);
@@ -220,40 +218,6 @@ export const SettingsSection: FC<SettingsSectionProps> = ({
         {internalSettings}
       </div>
     </div>
-  );
-};
-
-const AdvancedSettingsMessage = ({ shouldShowMessage }: { shouldShowMessage: boolean }) => {
-  const intl = useIntl();
-
-  const IntlMessage = {
-    message: intl.formatMessage({
-      defaultMessage:
-        'Add advanced parameters to further customize how the language model responds. These advanced options are optional and can help you fine-tune the experience.',
-      description: 'Agent advance parameters description',
-      id: 'yJfhNL',
-    }),
-    messageLink: intl.formatMessage({
-      defaultMessage: 'Guideline for agent advanced parameters',
-      description: 'Agent advance parameters description link',
-      id: '6F4Knf',
-    }),
-  };
-
-  return (
-    shouldShowMessage && (
-      <div style={{ paddingBottom: '5px' }}>
-        <Text style={{ fontSize: 12 }}>{IntlMessage.message} </Text>
-        <Link
-          href="https://learn.microsoft.com/en-us/azure/logic-apps/connectors/built-in/reference/openai/#get-chat-completions"
-          target="_blank"
-          style={{ fontSize: 12, fontStyle: 'italic' }}
-        >
-          {IntlMessage.messageLink}
-          <NavigateIcon style={{ position: 'relative', top: '2px', left: '2px' }} />
-        </Link>
-      </div>
-    )
   );
 };
